feat: [consolidate-inline-worker-clone-ownership-in-runtime-config-002] - [Lock inline runtime worker detachment behavior with focused runtime-config tests]
